Refactor resolveRef to perserve components in OpenAPI Documents
This commit updates the resovleRef function ti maintain the integrity of the original OpenAPI document. Previously, the function removed resolved components from the document's `components` section, which led to issues in subsequent reuse of the document, especially in test environments where the same document was dereferenced multiple times.

The refacotring includes:
= Removing the logic that deletes components from the root document after resolution.
- Adjusting related functions to align with this change (UnitTests)

This change ensures that eash test or operation that involves dereferencing an OpenAPI document works with a complete and unaltered version of the document, thereby preventing errors related to missing components in later tests or operations. It enhances the reliability and repeatability of tests and operations that depend on the OpenAPI document.